Move NamespacePath into Domain
