+++ phpstan.neon
@@ -64,7 +64,7 @@ parameters:
             message: 'which case rule a name follows is NameKind (symbols) or MemberKind (class-like members); applying one is NameCase; other uses may be allowlisted here consciously'
             allowIn:
                 - src/Domain/NameCase.php
-                - src/Utility/NamespacePath.php # namespace paths are case-insensitive for every kind; not a symbol name
+                - src/Domain/NamespacePath.php # namespace paths are case-insensitive for every kind; not a symbol name
                 - src/Transport/MessageReader.php # HTTP header names
                 - src/Domain/Visibility.php # visibility keywords
                 - tests/*
