+++ src/Index/AutoloadFilesLocator.php
@@ -7,10 +7,10 @@ namespace Firehed\PhpLsp\Index;
 use Firehed\PhpLsp\Cache\Invalidatable;
 use Firehed\PhpLsp\Document\FileUri;
 use Firehed\PhpLsp\Domain\NameKind;
+use Firehed\PhpLsp\Domain\NamespacePath;
 use Firehed\PhpLsp\Domain\QualifiedName;
 use Firehed\PhpLsp\Knowledge\SymbolLocator;
 use Firehed\PhpLsp\Parser\ParserService;
-use Firehed\PhpLsp\Utility\NamespacePath;
 use PhpParser\Node;
 
 /**
